requirements.txt
wheel
Mastodon.py
SQLAlchemy
streamlit
pandas